+++ package/example/windows/flutter/generated_plugins.cmake
@@ -3,6 +3,7 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
+  flutter_meedu_videoplayer
   fullscreen_window
   media_kit_video
   screen_brightness_windows