+++ package/CHANGELOG.md
@@ -1,3 +1,31 @@
+## 4.0.9
+
+* screenManager
+  * edgetoedge now is systemUiMode and is using SystemUiMode enum
+  * Added hideSystemOverlay option to enable and disable hiding system overlays
+* MeeduPlayerController 
+  * Variable windows to desktopOrWeb 
+  * Added excludeFocus variable (controls if widgets inside videoplayer should get focus or not)
+  * Added enabledControls in which gesture controls can be disabled 
+    * Defaults
+      * this.escapeKeyCloseFullScreen = true,
+      * this.numPadDecimalKeyToggleFit = true,
+      * this.enterKeyOpensFullScreen = true,
+      * this.spaceKeyTogglePlay = true,
+      * this.volumeArrows = true,
+      * this.seekArrows = true,
+      * this.seekSwipes = true,
+      * this.volumeSwipes = true,
+      * this.brightnessSwipes = true,
+      * this.doubleTapToSeek = true,
+      * this.desktopDoubleTapToFullScreen = true
+  * Added showLogs to be able to show and hide package logs
+  * Added manageBrightness to make disable handling of brightness
+* Fixed replay
+* Updated License
+* Fixed ios simulator crash
+  
+
 ## 4.0.8
 
 * Fixed macos&ios errors and added steps to readme for it